requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
jinja2==3.1.4
python-multipart==0.0.9
httpx==0.27.0
python-dotenv==1.0.1
aiofiles==23.2.1
